+++ src/widgets/history/ui/book_page/List.tsx
@@ -1,8 +1,8 @@
 import { artworks } from '@entities/history';
 import type { PageSide } from '@features/history/model/types';
 
+import { BookPageTitle } from '../BookPageTitle';
 import '../styles/List.css';
-import { BookPageTitle } from './BookPageTitle';
 
 const midpoint = Math.ceil(artworks.length / 2);
 const leftItems = artworks.slice(0, midpoint);
